Added purge command and limited username length.
Things slow down a lot when there are too many messages too fast. Added a way to remove all previous messages. I intend to be able to specify how many messages to remove, but for now it does everything.

Username length was also unlimited, and when a user entered an invalid name they would not be told the reason it was rejected. Both of these have also been fixed.